+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.25 branch
 ### 2.25.0
+* the `components/jquery` package was marked as a conflicting dependency and can no longer be installed;
 * `assertSqlEndsWith()`, `assertSqlEndsNotWith()`, `deleteLog()` and `getLogFullPath()` methods provided by the
   `TestCase` class was deprecated and has now been removed.