+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 * added `Command::isVerbose()` method;
 * added `VENDOR` and `WWW_VENDOR` constants;
 * improved `TestCase::assertLogContains()` method;
+* the `CreatePluginsLinksCommand` is now deprecated and will be removed in a later release. Use instead
+  `PluginAssetsSymlinkCommand` (`bin/cake plugin assets symlink`);
 * `deleteLog()` and `getLogFullPath()` methods provided by the `TestCase` class are now deprecated and will be removed
   in a later release;
-* `assertSqlEndsWith()` and `assertSqlEndsNotWith()` methods provided by the `TestCase` class now deprecated and will be
-  removed in a later release. Use instead `assertStringEndsWith()` and `assertStringEndsNotWith()`;
+* `assertSqlEndsWith()` and `assertSqlEndsNotWith()` methods provided by the `TestCase` class are now deprecated and
+  will be removed in a later release. Use instead `assertStringEndsWith()` and `assertStringEndsNotWith()`;
 * `LibraryHelper::analytics()` method was deprecated and has now been removed;
 * suggests `axllent/jquery` and no longer `components/jquery`. `axllent/jquery` added to `MeTools.VendorLinks` configuration;
 * updated for php-tools 1.7.4.